requirements.txt
asgiref==3.3.1
Django==3.1.5
django-etc==1.3.0
django-filter==2.4.0
django-hitcount==1.3.3
django-search-views==0.3.7
django-taggit==1.3.0
mysqlclient==2.0.3
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
